chore(security): :lock: bump `micromatch` to version `^4.0.8` to address CVE-2024-4068 and updates some other dev dependencies
`micromatch` is an indirect dependency from `ts-loader`, and it also has a dependency called `braces`, which in all versions prior to `3.0.3` have the CVE-2024-4068 vulnerability. The `micromatch`'s maintainer(s) have addressed this issue on version `4.0.8`, but `ts-loader` still uses `micromatch@4.0.5`, hence this override of an indirect dependency. This new versions should not break as them only fix bugs and adresses this CVE-2024-4068. This commit also updates `webpack` to the latest version (`5.99.9`) as to address CVE-2024-43788, that was fixed since version `5.94.0`. Also updated other dependencies just to keep everything updated.